food-classification/pyproject.toml
[tool.poetry]
name = "food-classification"
version = "0.1.0"
description = ""
authors = ["Eric Xia"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.6.1"
torch = "^2.6.0"
tqdm = "^4.67.1"
pillow = "^11.1.0"
matplotlib = "^3.10.0"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
torchvision = "^0.21.0"
sentence-transformers = "^3.4.1"
sentencepiece = "^0.2.0"
datasets = "^3.3.2"
opencv-python = "^4.11.0.86"
captum = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
